refactor: 결제 클라이언트 설정 제거 및 결제 클라이언트 추가
